+++ telegras/backends/__init__.py
@@ -1,4 +1,3 @@
-from .wordpress import WordPressPublishBackend
 from .factory import build_backends, parse_backend_names
 
-__all__ = ["WordPressPublishBackend", "build_backends", "parse_backend_names"]
+__all__ = ["build_backends", "parse_backend_names"]